# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Transferências e Usuários",
    "version": "1.0.0",
    "description": "API para registro, login, consulta de usuários e transferências."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [],
  "paths": {
    "/api/users/register": {
      "post": {
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "favorecido": { "type": "boolean" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário criado" },
          "409": { "description": "Usuário já existe" },
          "400": { "description": "Dados inválidos" }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "summary": "Login de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login realizado" },
          "401": { "description": "Login ou senha inválidos" },
          "400": { "description": "Dados inválidos" }
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "Listar usuários",
        "responses": {
          "200": { "description": "Lista de usuários" }
        }
      }
    },
    "/api/transfers": {
      "post": {
        "summary": "Realizar transferência",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "from": { "type": "string" },
                  "to": { "type": "string" },
                  "amount": { "type": "number" }
                },
                "required": ["from", "to", "amount"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Transferência realizada" },
          "400": { "description": "Erro na transferência" }
        }
      },
      "get": {
        "summary": "Listar transferências",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Lista de transferências" }
        }
      }
    }
  }
}